Refactor: Estandarización visual de listados y unificación de experiencia de usuario (UI/UX)
- UI: Se reemplazaron los Spinners clásicos por Material Design Exposed Dropdown Menus en todos los filtros de Inventario.
- UI: Unificación de Toolbars con buscador integrado en vistas de Operario (Inventario y Maquinaria).
- UI: Eliminación de filtros de estado (Activo/Inactivo) en inventario de operarios para simplificar la vista y ocultar productos no operativos.
- Design: Ajuste de márgenes y paddings en la vista de Usuarios para cumplir con el estándar de 28dp de la aplicación.
- UX: Eliminación de cabeceras informativas redundantes y simplificación de layouts en módulos de operario.
- Fix: Corrección de errores de compilación por referencias a IDs de ViewBinding inexistentes en InventarioOpFragment.
- Fix: Estandarización de botones de acción (Extended FAB) eliminando iconos duplicados y unificando textos.

+++ app/src/main/java/com/lingomak/lingomakapp/ui/dashboard/home/HomeOperarioFragment.kt
@@ -103,8 +103,8 @@ class HomeOperarioFragment : Fragment() {
                         putString("codigoMaquinaria", maquinaria.codigoMaquinaria)
                         putString("nombreMaquinaria", maquinaria.nombre)
                         putString("tipoMaquinaria", maquinaria.tipo)
-                        putInt("horometroActual", maquinaria.horometroActual)
-                        putInt("horometroUltimoMantenimiento", maquinaria.horometroUltimoMantenimiento)
+                        putDouble("horometroActual", maquinaria.horometroActual)
+                        putDouble("horometroUltimoMantenimiento", maquinaria.horometroUltimoMantenimiento)
                         putInt("intervaloMantenimientoHoras", maquinaria.intervaloMantenimientoHoras)
                     }
                     fragment.arguments = bundle
